MAINT: split package to scripts
