+++ dtoolkit/accessor/series/__init__.py
@@ -1,0 +1,9 @@
+from dtoolkit.accessor.series.bin import bin  # noqa
+from dtoolkit.accessor.series.cols import cols  # noqa
+from dtoolkit.accessor.series.drop_inf import drop_inf  # noqa
+from dtoolkit.accessor.series.error_report import error_report  # noqa
+from dtoolkit.accessor.series.expand import expand  # noqa
+from dtoolkit.accessor.series.getattr import get_attr  # noqa
+from dtoolkit.accessor.series.len import lens  # noqa
+from dtoolkit.accessor.series.top_n import top_n  # noqa
+from dtoolkit.accessor.series.values_to_dict import values_to_dict  # noqa
